Echidna tests and tooling updates (#552)
* Include echidna contracts in formatting check

* Add solium rules

* Add slither script error message

* Restructure test directories

* Add new echidna tests

* Remove vault tests

* Remove ENSBuffer.sol and fix Echidna tests

* Regenerate contract bindings

* Improve echidna prefix and test names

* Update node indexes in CI

+++ slither/slither.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
 
+if [ ! -d "contracts" ]; then 
+	echo "error: script needs to be run from project root './slither/slither.sh'"
+	exit 1
+fi
+
 docker run --rm -v "$PWD":/contracts -it --workdir=/contracts/slither trailofbits/eth-security-toolbox@sha256:3fb96e2d9de772f5e97f1c3c650c8a3d28660f8a64a60b76269da1ac19b86a28 -c 'solc-select 0.5.17 && slither --triage-mode ../contracts'
